requirements.txt
sqlalchemy>=2.0.42,<3.0.0
sqlmodel>=0.0.24,<0.0.25
fastapi[standard]>=0.116.1,<0.117.0
python-dotenv>=0.9.9,<0.10.0
psycopg2>=2.9.10,<3.0.0
uvicorn[standard]
